requirements.txt
# Core scientific computing
numpy>=1.21.0
scipy>=1.7.0
matplotlib>=3.5.0
scikit-learn>=1.0.0

# Machine Learning (macOS ARM64 compatible)
tensorflow-macos>=2.8.0
tensorflow-metal>=0.1.0
jax>=0.3.0
jaxlib>=0.3.0

# Quantum Computing
pennylane>=0.25.0
pennylane-lightning>=0.25.0

# Data handling
h5py>=3.6.0
